Split osu web services by domain
